Time: 3 ms (85.52%) | Memory: 16.3 MB (12.65%) - LeetSync

+++ 152-maximum-product-subarray/maximum-product-subarray.cpp
@@ -1,9 +1,9 @@
 class Solution {
 public:
     int maxProduct(vector<int>& nums) {
-        int maxi = INT_MIN;
-        int p = 1;
-        int s = 1;
+        double  maxi = INT_MIN;
+        double p = 1;
+        double s = 1;
         int n = nums.size();
         for (int i = 0;i<nums.size();i++){
             if (p==0){
